Refactor entity mapping and update namespaces
In `ApplicationMappingProfile.cs`, added namespaces `Assignment.Application.TodoLists.Queries.GetTodos` and `Assignment.Domain.Entities`. Created a new class `ApplicationMappingProfile` inheriting from `Profile` and defined mappings for `TodoList` to `TodoListDto` and `TodoItem` to `TodoItemDto`. In `LookupDto.cs`, removed the namespace `Assignment.Domain.Entities` and the private class `Mapping` that was previously responsible for mapping `TodoList` and `TodoItem` to `LookupDto`. This indicates a shift in responsibility for these mappings to the `ApplicationMappingProfile` class.