Added MaterialDesignThemes and Snackbar functionality to project
MaterialDesignThemes package version 5.0.0 was added to the project, providing Material Design styles for WPF applications. The Snackbar control was added to TodoListView.xaml for displaying brief messages. SnackbarMessageQueue property was added to TodoListViewModel.cs to enqueue messages for Snackbar. The MessageBox.Show method call in SaveExecute was replaced with Snackbar's Enqueue method. The layout of TodoListView.xaml was changed from StackPanel to Grid for flexible control positioning. MaxLengthRule validation rule was commented out, removing the maximum length for the Title property. ISnackbarMessageQueue interface was added to TodoListViewModel.cs for enqueuing Snackbar messages.